.update how we create context
