+++ apps/ow-api/src/trpc.ts
@@ -1,13 +1,15 @@
-import { getLogger } from "@dotkomonline/logger"
 import { z } from "zod"
 
 import { initTRPC } from "@trpc/server"
 
-import { Context } from "./index.js"
+import { Context } from "./context"
 
-const logger = getLogger(import.meta.url)
-
-export const t = initTRPC.context<Context>().create()
+// TODO: Superjson
+export const t = initTRPC.context<Context>().create({
+  errorFormatter({ shape }) {
+    return shape
+  },
+})
 
 export const appRouter = t.router({
   signin: t.procedure
@@ -19,7 +21,7 @@ export const appRouter = t.router({
       })
     )
     .mutation((req) => {
-      logger.info(JSON.stringify(req))
+      console.log("XD")
       return { msg: "hello world" }
     }),
 })
