+++ apps/web/package.json
@@ -46,6 +46,8 @@
   },
   "devDependencies": {
     "@dotkomonline/config": "workspace:*",
+    "@dotkomonline/db": "workspace:^",
+    "@dotkomonline/logger": "workspace:^",
     "@storybook/addon-actions": "^6.5.9",
     "@storybook/addon-essentials": "^6.5.9",
     "@storybook/addon-knobs": "^6.4.0",
