+++ apps/web/src/utils/trpc.ts
@@ -1,6 +1,6 @@
 // utils/trpc.ts
 import { createTRPCReact } from "@trpc/react-query"
 
-import type { AppRouter } from "../../../ow-api/src/trpc"
+import { type AppRouter } from "../../../ow-api/src/trpc"
 
 export const trpc = createTRPCReact<AppRouter>()
